added aws_utils folder, refactored accordingly
