+++ src/app/command_handler.py
@@ -2,7 +2,7 @@
 from flask import json, jsonify
 import requests
 from config import GOOGLE_BOOKS_API_URL
-from dynamodb import get_current_book
+from aws_utils.dynamodb import get_current_book
 from utils import random_greeting
 
 def command_handler(raw_request, current_books_list):
