+++ src/app/helper_functions.py
@@ -1,6 +1,6 @@
 from flask import jsonify
 from utils import is_valid_future_date
-from dynamodb import put_book, get_current_book
+from aws_utils.dynamodb import put_book, get_current_book
 
 
 def handle_book_select(raw_request, current_books_list, pending_selections):
